refactor: extract websocket provider
